Move classes into separate files to simplify code layout
